+++ CHANGELOG.md
@@ -8,7 +8,8 @@
 
 ### Bug fixes
 
-* Invoke callbacks in the right order: `A.Vector.map/2`
+* Invoke callbacks in the right order: `A.Vector.map/2`,
+  `A.Vector.map_intersperse/3`, `A.Vector.map_join/3`
 
 ## v0.4.1 (2020-12-05)